Change DataType::FixedSize*(i32) to DataType::FixedSize*(usize)
Fixes #525

+++ src/io/avro/read/schema.rs
@@ -173,7 +173,7 @@ fn schema_to_field(
                 false,
             ))
         }
-        AvroSchema::Fixed { size, .. } => DataType::FixedSizeBinary(*size as i32),
+        AvroSchema::Fixed { size, .. } => DataType::FixedSizeBinary(*size),
         AvroSchema::Decimal {
             precision, scale, ..
         } => DataType::Decimal(*precision, *scale),
